pyproject.toml
[project]
name = "arc-agi3"
version = "0.1.0"
description = "Development environment for the ARC-AGI-3 interactive reasoning benchmark"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "accelerate>=1.12.0",
    "anthropic>=0.77.0",
    "arc-agi>=0.9.1",
    "llama-cpp-python>=0.3.16",
    "ollama>=0.6.1",
    "peft>=0.18.1",
    "pytest>=9.0.2",
    "python-dotenv>=1.2.1",
    "transformers>=5.1.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.3.0",
    "pre-commit>=3.6.0",
]
agentops = [
    "agentops>=0.3.0",
]
llm = [
    "openai>=1.0.0",
    "anthropic>=0.18.0",
]
aria-lite = [
    "torch>=2.0.0",
    "llama-cpp-python>=0.2.0",
]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
ignore = ["E501"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
